Use the `proc_macro` token model in the bootstrap grammar.
