+++ build.rs
@@ -9,6 +9,8 @@ extern crate proc_quote;
 mod generate;
 #[path = "src/parse_node.rs"]
 mod parse_node;
+#[path = "src/proc_macro.rs"]
+pub mod proc_macro;
 #[path = "src/scannerless.rs"]
 pub mod scannerless;
 
@@ -16,14 +18,13 @@ use std::env;
 use std::fs;
 use std::path::PathBuf;
 
-// FIXME(eddyb) use `scannerless::Grammar` when that wrapper hack is fixed.
-type Grammar = grammer::Grammar<scannerless::Pat<&'static str>>;
 fn main() {
     println!("cargo:rerun-if-changed=build.rs");
 
     let out_dir = PathBuf::from(env::var("OUT_DIR").unwrap());
 
-    let grammar: Grammar = grammer::grammar_grammar();
+    let mut grammar = proc_macro::builtin();
+    grammar.extend(grammer::grammar_grammar());
 
     fs::write(
         &out_dir.join("parse_grammar.rs"),
