+++ macros/src/lib.rs
@@ -7,28 +7,16 @@ use proc_quote::ToTokens as _;
 
 #[proc_macro]
 pub fn scannerless_parser(input: TokenStream) -> TokenStream {
-    // FIXME(eddyb) parse the `proc_macro` tokens instead of strings.
-    gll::generate::rust::generate(
-        &input
-            .to_string()
-            .parse::<gll::scannerless::Grammar>()
-            .unwrap(),
-    )
-    .into_token_stream()
-    .into()
+    let grammar: gll::scannerless::Grammar = gll::parse_grammar(input.into()).unwrap();
+    gll::generate::rust::generate(&grammar)
+        .into_token_stream()
+        .into()
 }
 
 #[proc_macro]
 pub fn proc_macro_parser(input: TokenStream) -> TokenStream {
-    // FIXME(eddyb) parse the `proc_macro` tokens instead of strings.
     let mut grammar = gll::proc_macro::builtin();
-    grammar.extend(
-        input
-            .to_string()
-            .parse::<gll::proc_macro::Grammar>()
-            .unwrap()
-            .0,
-    );
+    grammar.extend(gll::parse_grammar(input.into()).unwrap());
     gll::generate::rust::generate(&grammar)
         .into_token_stream()
         .into()
